requirements.txt
wheel
langchain
PyPDF2
python-dotenv
streamlit
streamlit-extras
openai
faiss-cpu
wikipedia
duckduckgo-search
cohere
tiktoken
boto3
shodan
# sklearn replaced by scikit-learn
#sklearn
scikit-learn
matplotlib
umap-learn
#kedra stuff
tabulate
requests
# uncomment to use huggingface llms
# huggingface-hub

# uncomment to use instructor embeddings
# InstructorEmbedding
# sentence-transformers

# GraphQL tool deps
httpx
gql
requests_toolbelt

# Howler tool deps
howler_client

trino
trino[sqlalchemy]